refactor: extract shared periodOrder/collapseContiguousPeriods into CourseScheduleUtils
Removes duplicate implementations from LiveActivityResolver and ClassPreparingNotificationScheduler, both now delegate to the single canonical helpers in the data package.
